fix: revert custom llm functions
